Chore: Refatora - models do Authors
