+++ T19C-Back-End/bloco-23-NodeJs-Camada-de-Servico-e-Arquitetura-Rest-e-Restful/dia-02-Arquitetura-de-Software-Camada-de-Controller-e-Service/src/para-fixar/model-example/models/Author.js
@@ -46,14 +46,6 @@ const findById = async (id) => {
   });
 };
 
-const isValid = (firstName, middleName, lastName) => {
-  if (!firstName || typeof firstName !== 'string') return false;
-  if (!middleName || typeof middleName !== 'string') return false;
-  if (!lastName || typeof lastName !== 'string') return false;
-
-  return true;
-};
-
 const create = async (firstName, middleName, lastName) => connection.execute(
   'INSERT INTO model_example.authors (first_name, middle_name, last_name) VALUES (?,?,?)',
   [firstName, middleName, lastName],
@@ -62,6 +54,5 @@ const create = async (firstName, middleName, lastName) => connection.execute(
 module.exports = {
   getAll,
   findById,
-  isValid,
   create,
 };
