+++ T19C-Back-End/bloco-23-NodeJs-Camada-de-Servico-e-Arquitetura-Rest-e-Restful/dia-02-Arquitetura-de-Software-Camada-de-Controller-e-Service/src/para-fixar/model-example/services/Book.js
@@ -1,6 +1,5 @@
 // services/Book.js
 
-const Author = require('../models/Author');
 const Book = require('../models/Books');
 
 const getAll = async () => Book.getAll();
@@ -20,23 +19,7 @@ const findById = async (id) => {
   return book;
 };
 
-const createBook = async (title, bookId) => {
-  const author = await Author.findById(bookId);
-
-  if (!author) {
-    return {
-      error: {
-        code: 'notFound',
-        message: 'Autor não encontrado',
-      },
-    };
-  }
-
-  return Book.createBook(title, bookId);
-};
-
 module.exports = {
   getAll,
   findById,
-  createBook,
 };
